\name{nfca_example}
\docType{data}
\alias{nfca_example}
\title{Family Correlation Coefficient Data}
\description{
  This data is correlation coefficient matrix among family members. The family members are Husband(H), Wife(W), Son(S), Daughter(D), Husband's Father(HF), 
  Husband's Mother(HM), Wife's Father(WF), Wife's Mother(WM), Husband's Friend(H1), Wife's Friend(W1), and Others(O).
}
\usage{nfca_example}
\format{data frame}
\source{Ma, J. (2010), Contributions to Numerical Formal Concept Analysis, Bayesian Predictive Inference, and Sample Size Determination, PhD thesis, 
Case Western Reserve University. \cr
http://rave.ohiolink.edu/etdc/view?acc_num=case1285341426}
\references{
Ma, J., Sun, J., and Zhang, G.-Q. (2014), Numerical Formal Concept Analysis (nFCA): a New Systematic Clustering Technique. Under review.
}
\keyword{datasets}